feat: split C file into different files (#19)
